requirements.txt
asgiref             
bencode.py         
Django               
django-crispy-forms           
Pillow                                
setuptools          
sqlparse             
wheel
mysqlclient                
